==> step0_prepare_data.ipynb <==
# !pip install --user pandas-profiling

import os

import numpy as np
import pandas as pd
import pandas_profiling
import datetime

!mkdir -p data

RAW_DATA_PATH = './raw_data/'
DATA_PATH = './data'

!ls raw_data/

raw_data_likes = pd.read_csv(os.path.join(RAW_DATA_PATH, 'train_likes.csv'))
raw_data_likes['day'] = pd.to_datetime(raw_data_likes['day'])
raw_data_likes.head()

def get_user_day_ordinal(df):
    return np.concatenate(
        (
            df
                .groupby(['user_id', 'day'] , sort=False)['day']
                .apply(lambda s: np.arange(len(s)))
        ).tolist()
    )

def get_pseudo_ts(df):
    return np.concatenate(
        (
            df
                .groupby(['user_id', 'day'], sort=False)['day']
                .apply(lambda s: s.min() +  ((datetime.timedelta(days=1) / len(s)) * np.arange(len(s))))
        ).tolist()
    )

# def filter_users_pictures(df):
#     df_user_actions_count = df.groupby('user_id')['picture_id'].count()
#     selected_users = df_user_action_count[df_user_actions_count > 1].index.tolist()
#     df = df[df['user_id'].isin(selected_users)]
    
#     df_user_actions_count = cutted_train_clicks.groupby('user_id')['picture_id'].count()
#     selected_users = df_user_action_count[df_user_actions_count > 1].index.tolist()
#     df = df[df['user_id'].isin(selected_users)]
    
#     df_picture_actions_count = df.groupby('picture_id')['user_id'].count()
#     selected_pictures = df_picture_actions_count[df_picture_actions_count > 1].index.tolist()
#     df = df[df['picture_id'].isin(selected_pictures)]
    
#     return df

##### Likes

raw_data_likes['user_day_ordinal'] = get_user_day_ordinal(raw_data_likes)
raw_data_likes['pseudo_ts'] = get_pseudo_ts(raw_data_likes)
raw_data_likes.sort_values(['user_id', 'day']).head(10)

raw_data_likes.to_hdf(os.path.join(DATA_PATH, 'step0_likes.h5'), index=None, key='step0')

all(new_raw_data_likes == raw_data_likes)

##### Clicks

raw_data_clicks = pd.read_csv(os.path.join(RAW_DATA_PATH, 'train_clicks.csv'))
raw_data_clicks['day'] = pd.to_datetime(raw_data_clicks['day'])
raw_data_clicks.head()

raw_data_clicks['user_day_ordinal'] = get_user_day_ordinal(raw_data_clicks)
raw_data_clicks['pseudo_ts'] = get_pseudo_ts(raw_data_clicks)
raw_data_clicks.sort_values(['user_id', 'day'], ascending=False).tail(10)

raw_data_clicks.to_hdf(os.path.join(DATA_PATH, 'step0_clicks.h5'), index=None, key='step0')

##### Shares

raw_data_shares = pd.read_csv(os.path.join(RAW_DATA_PATH, 'train_shares.csv'))
raw_data_shares['day'] = pd.to_datetime(raw_data_shares['day'])
raw_data_shares.head()

raw_data_shares['user_day_ordinal'] = get_user_day_ordinal(raw_data_shares)
raw_data_shares['pseudo_ts'] = get_pseudo_ts(raw_data_shares)
raw_data_shares.sort_values(['user_id', 'day']).tail(10)

raw_data_shares.to_hdf(os.path.join(DATA_PATH, 'step0_shares.h5'), index=None, key='step0')

##### test_users

test_users = pd.read_csv(os.path.join(RAW_DATA_PATH, 'test_users.csv'))
test_users.head()

# Mapping

##### User

# user mapping
all_user_ids = np.sort(np.unique(
    raw_data_likes['user_id'].unique().tolist() +
    raw_data_clicks['user_id'].unique().tolist() +
    raw_data_shares['user_id'].unique().tolist() +
    test_users['user_id'].unique().tolist()
))

user_mapping = pd.DataFrame({'old': all_user_ids, 'new': np.arange(len(all_user_ids))})

user_mapping.to_csv(os.path.join(DATA_PATH, 'step0_user_mapping.csv'), index=False)

##### Pic

# picture mapping
all_picture_ids = np.sort(np.unique(
    raw_data_likes['picture_id'].unique().tolist() +
    raw_data_clicks['picture_id'].unique().tolist() +
    raw_data_shares['picture_id'].unique().tolist() +
    raw_picture_descriptions['picture_id'].unique().tolist()
))

picture_mapping = pd.DataFrame({'old': all_picture_ids, 'new': np.arange(len(all_picture_ids))})

len(set(raw_picture_descriptions['picture_id'].unique().tolist()) & set(picture_mapping['old'].unique().tolist()))

picture_mapping.to_csv(os.path.join(DATA_PATH, 'step0_picture_mapping.csv'), index=False)

# Descriptions

raw_picture_descriptions = pd.read_csv(os.path.join(RAW_DATA_PATH, 'descriptions.csv'))
raw_picture_descriptions.head()

# attribute mapping
pictures = []
descs = []
for picture_id, picture_descs in raw_picture_descriptions[['picture_id', 'description']].values:
    pictures += np.repeat([picture_id], repeats=len(picture_descs)).tolist()
    descs += picture_descs

picture_descs_df = pd.DataFrame({'picture_id': pictures, 'description_uid': descs})

grouper_attr = picture_descs_df.groupby('description_uid').grouper
desc_mapping = pd.DataFrame({'new': grouper_attr.group_info[1], 'old': grouper_attr.levels[0]})

desc_mapping.to_csv(os.path.join(DATA_PATH, 'desc_mapping.csv'), index=False)

!ls -alh data/
